evo2: 7b/40b configs rotary embedding base (#171)
* evo2: 40b: Fix rotary embedding base

The issue is currently benign as inv_freq (which is derived from base)
is loaded from the checkpoint (as `self.inv_freq` is registered as a
buffer.)

It is still good to fix it from correctness (and documentation) point of
view.

```
>>> def inv_freq(base, dim): return 1.0 / (base ** (torch.arange(0, dim, 2, device="cuda", dtype=torch.float32) / dim))

>>> pt = torch.load("evo2_40b.pt", weights_only=False, mmap=True, map_location="cpu")

>>> torch.mean(pt["blocks.3.inner_mha_cls.rotary_emb.inv_freq"] - inv_freq(1e11, 128).cpu())
tensor(0.0326)

>>> torch.mean(pt["blocks.3.inner_mha_cls.rotary_emb.inv_freq"] - inv_freq(1e6, 128).cpu())
tensor(-2.5294e-05)
```

* evo2: 7b: Fix rotary embedding base

The issue is currently benign as inv_freq (which is derived from base)
is loaded from the checkpoint (as `self.inv_freq` is registered as a
buffer.)

It is still good to fix it from correctness (and documentation) point of
view.

```
>>> def inv_freq(base, dim): return 1.0 / (base ** (torch.arange(0, dim, 2, device="cuda", dtype=torch.float32) / dim))

>>> pt = torch.load("evo2_7b.pt", weights_only=False, mmap=True, map_location="cpu")

>>> torch.mean(pt["blocks.3.inner_mha_cls.rotary_emb.inv_freq"] - inv_freq(1e11, 128).cpu())
tensor(0.0688)

>>> torch.mean(pt["blocks.3.inner_mha_cls.rotary_emb.inv_freq"] - inv_freq(1e6, 128).cpu())
tensor(0.0361)

>>> torch.mean(pt["blocks.3.inner_mha_cls.rotary_emb.inv_freq"] - inv_freq(1e4, 128).cpu())
tensor(5.2014e-06)
```